Update
Moved dispose handling to NdisApiUser
Now all of the forms exit gracefully (OnClosed event is fired on exit)
TorService and ProxiFyreService are now trying to exit gracefully
Fixed bug with ProxiFyre: adapter mode is not switching back from tunnel mode to 0

+++ TorCSClient/TorCSClient/Proxy/TorService.cs
@@ -246,7 +246,10 @@ namespace TorCSClient.Proxy
             {
                 //_TorCSClientProcess?.StandardInput.Close();
                 //_TorCSClientProcess?.StandardOutput.Close();
-                _torProcess?.Kill(entireProcessTree: true);
+                if (!_torProcess.CloseMainWindow())
+                {
+                    _torProcess?.Kill(entireProcessTree: true);
+                }
                 //_TorCSClientProcess?.WaitForExit();
             }
             WorkingBridges.Clear();
